refactor(shared): extract resolve_mcp_tools/cleanup_mcp_clients into shared.mcp_tools
MCP tool resolution is shared by agent loops that need server:mcp tool IDs. Centralize resolution and cleanup while keeping MCP code out of plugins that do not use it.
